second commit

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth')->only('store', 'destroy');
+    }
+
     public function index() {
 
         $posts = Post::orderBy('created_at', 'desc')->with('user', 'likes')->paginate(5);
@@ -31,7 +36,15 @@ class PostController extends Controller
         return back();
     }
 
+    public function show(Post $post) {
+        return view('posts.show', [
+            'post' => $post
+        ]);
+    }
+
     public function destroy(Post $post) {
+
+        $this->authorize('delete', $post);
 
         $post->delete();
